requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.39.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.31.0
